pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = [
  "poetry-core",
]

[tool.poetry]
name = "toolbox"
version = "0.1.0"
description = "Toolbox aims to be one of the best devops helper."
license = "MIT"
authors = [
  "tyvekZhang <tyvek.zhang@gmail.com>",
]
maintainers = [
  "tyvekZhang <tyvek.zhang@gmail.com>",
]
readme = "README.md"
homepage = "https://github.com/dylenchang/toolbox"
repository = "https://github.com/dylenchang/toolbox"
documentation = "https://github.com/dylenchang/toolbox/wiki"
keywords = ["fastapi", "sqlmodel", "tools", "web", "scaffold"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
packages = [
  { include = "toolbox"},
]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.2"
fastapi = "^0.110.0"
fastapi-pagination = "^0.12.21"
fastapi-async-sqlalchemy = "^0.6.1"
fastapi-offline = "^1.7.1"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
bcrypt = "==4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
alembic = "^1.13.1"
sqlalchemy-utils = "^0.41.2"
aiosqlite = "^0.20.0"
eval-type-backport = "^0.1.3"
diskcache = "^5.6.3"
pandas = "^2.2.2"
xlsxwriter = "^3.2.0"
openpyxl = "^3.1.2"
slowapi = "^0.1.9"
asyncpg = "^0.29.0"
aiomysql = "^0.2.0"
redis = {extras = ["hiredis"], version = "^5.0.3"}
greenlet = "^3.0.3"
websocket = "^0.2.1"
websockets = "^12.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
coverage = "^7.4.4"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"

[tool.poetry.scripts]
[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.coverage.run]
parallel = true
branch = true
source = ["src/common", "src/middleware", "src/starter"]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    ]
